my_functions moved to io_functions (input-output)
