+++ macros/Average.py
@@ -1,6 +1,6 @@
 import sys
 sys.path.insert(0, '../')
-from lib.my_functions import load_npy
+from lib.io_functions import load_npy
 from lib.wvf_functions import average_wvfs
 
 N_runs     =[10,22,26]
